Add trace data to trace_2.txt for memory read operations
This commit introduces a new file, trace_2.txt, containing a series of memory read (R) operations with various addresses. The data is intended for tracking and analyzing memory access patterns.

Note that trace file can be read by 8 way set associative cache as of now. It can be extended to other cache as well.